+++ src/main/java/nextstep/reservation/dao/ReservationWaitingDao.java
@@ -1,9 +1,10 @@
-package nextstep.reservation;
+package nextstep.reservation.dao;
 
 import java.sql.PreparedStatement;
 import java.util.Collections;
 import java.util.List;
 import nextstep.member.Member;
+import nextstep.reservation.domain.ReservationWaiting;
 import nextstep.schedule.Schedule;
 import nextstep.theme.Theme;
 import org.springframework.jdbc.core.JdbcTemplate;
