+++ src/main/java/nextstep/reservation/domain/Reservation.java
@@ -1,25 +1,26 @@
-package nextstep.reservation;
+package nextstep.reservation.domain;
 
 import java.util.Objects;
 import nextstep.member.Member;
 import nextstep.schedule.Schedule;
 
-public class ReservationWaiting {
+public class Reservation {
 
-    public Long id;
-    public Schedule schedule;
-    public Member member;
-    public Integer waitNum;
+    private Long id;
+    private Schedule schedule;
+    private Member member;
 
-    public ReservationWaiting(Schedule schedule, Member member, Integer waitNum) {
-        this(null, schedule, member, waitNum);
+    public Reservation() {
     }
 
-    public ReservationWaiting(Long id, Schedule schedule, Member member, Integer waitNum) {
+    public Reservation(Schedule schedule, Member member) {
+        this(null, schedule, member);
+    }
+
+    public Reservation(Long id, Schedule schedule, Member member) {
         this.id = id;
         this.schedule = schedule;
         this.member = member;
-        this.waitNum = waitNum;
     }
 
     public Long getId() {
@@ -32,10 +33,6 @@ public class ReservationWaiting {
 
     public Member getMember() {
         return member;
-    }
-
-    public Integer getWaitNum() {
-        return waitNum;
     }
 
     public boolean sameMember(Member member) {
